requirements.txt


# Núcleo
requests>=2.28.0
requests-cache>=0.9.8
pandas>=2.0.0
numpy>=1.23.0
scikit-learn>=1.2.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# ML Avanzado (opcionales; el código funciona sin ellos)
xgboost>=2.0.0
lightgbm>=4.0.0

# Visualización (opcional)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Series temporales (opcional)
statsmodels>=0.14.0

# Multi-fuente (opcionales; solo si activarás IMERG/GPCP)
xarray>=2023.1.0
netCDF4>=1.6.0
h5py>=3.8.0
pydap>=3.4.0
earthaccess>=0.8.0

# Utilidades
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0
